test: Disable hostfw only for kube-proxy + externalTrafficPolicy=Local
For the externalTrafficPolicy=Local test with kube-proxy, we disable the
host firewall because of #12542. That leads us to reinstalling Cilium
with the host firewall explictly disabled.

We however don't reinstall Cilium afterwards with the default configs
(including potentially the host firewall). This commit fixes that. We
also don't need to two Cilium reinstallation if the host firewall is
already disabled.

Signed-off-by: Paul Chaignon <paul@cilium.io>